Basic Returning hours to user on cancelling setup from myBookings page

+++ src/php/bookings/cancel.php
@@ -32,11 +32,84 @@
 		http_response_code(406); //Invalid Entry
 	} else {
 		//Cancel booking
-		$sth = $db->prepare("DELETE FROM BookingSlots WHERE bookingID = ?");
-		$sth->execute(array($bookingID));
-		$sth = $db->prepare("DELETE FROM Bookings WHERE bookingID = ?");
-		$sth->execute(array($bookingID));
-		http_response_code(200);		
+
+		//if user is not admin, confirm they are trying to cancel their own booking
+		if ($_SESSION["class"] != "admin"){
+			$sth = $db->prepare("SELECT uID FROM Bookings WHERE bookingID = ?");
+			$sth->execute(array($bookingID));
+			while($row = $sth->fetch(PDO::FETCH_ASSOC)) {
+				$userID = $row['uID'];
+			}
+			
+			//confirm this is the user's booking
+			if($userID != $_SESSION["netID"]){
+				http_response_code(406); //Invalid Entry
+			}
+			else{
+				//determine amount of time booking was
+
+				$sth = $db->prepare("SELECT bookingDate, COUNT(*) as numBlocks FROM BookingSlots WHERE bookingID = ?");
+				$sth->execute(array($bookingID));
+				
+				while($row = $sth->fetch(PDO::FETCH_ASSOC)) {
+					//Get number of blocks
+					$numBlocks = $row['numBlocks'];
+					$bookingDate = $row['bookingDate'];
+				}
+				$bookingLength = ($numBlocks*30.0)/60.0; //hours to be returned to the user
+				echo $bookingLength;
+
+				//cancel booking
+				$sth = $db->prepare("DELETE FROM BookingSlots WHERE bookingID = ?");
+				$sth->execute(array($bookingID));
+				$sth = $db->prepare("DELETE FROM Bookings WHERE bookingID = ?");
+				$sth->execute(array($bookingID));
+
+				//if today is Sunday, then must use Monday last week to retrieve hours for the current week
+				//TODO
+
+				//determine current date so you can determine where to return hours
+				$firstDay = date("Y-m-d", strtotime('monday this week'));  
+				$firstDayNextWeek = date("Y-m-d", strtotime('monday next week'));
+				$firstDayWeek3 = date("Y-m-d", strtotime('monday next week next week'));  
+
+				//if booking made in the current week
+				if($bookingDate >= $firstDay && $bookingDate < $firstDayNextWeek) {
+					$week = 'curWeekHrs';
+
+				} else if($bookingDate < $firstDayWeek3)  {
+				    $week = 'nextWeekHrs';
+				} 
+				else{
+					$week ='thirdWeekHrs';
+				}
+				echo $week;
+
+				//determine the number of hours the user currently has for that week
+				$sth = $db->prepare("SELECT $week FROM User WHERE uID = ?");
+				$sth->execute(array($userID));
+				while($row = $sth->fetch(PDO::FETCH_ASSOC)) {
+					//Get number of blocks
+					$bookingLength = $row[$week] + $bookingLength;
+				}
+				echo $bookingLength;
+
+				//update the user table with the retruned hours from the canceled booking
+				$sth = $db->prepare("UPDATE User SET $week = ? WHERE uID = ?");
+				$sth->execute(array($bookingLength,$userID));
+
+				http_response_code(200); //sucess
+			}
+
+		}
+		else{ //user is admin
+			$sth = $db->prepare("DELETE FROM BookingSlots WHERE bookingID = ?");
+			$sth->execute(array($bookingID));
+			$sth = $db->prepare("DELETE FROM Bookings WHERE bookingID = ?");
+			$sth->execute(array($bookingID));
+			http_response_code(200); //sucess
+		}
+		
 	}
 
 	//Close the connection
